+++ integration-test/integration/initialize_test.clj
@@ -12,11 +12,10 @@
 (deftest default-initialize-and-shutdown
   (eca/start-process!)
 
-  (let [models ["anthropic/claude-haiku-4-5-20251001"
-                "anthropic/claude-opus-4-1-20250805"
-                "anthropic/claude-opus-4-20250514"
-                "anthropic/claude-sonnet-4-20250514"
-                "anthropic/claude-sonnet-4-5-20250929"
+  (let [models ["anthropic/claude-haiku-4.5"
+                "anthropic/claude-opus-4.1"
+                "anthropic/claude-opus-4.5"
+                "anthropic/claude-sonnet-4.5"
                 "github-copilot/claude-haiku-4.5"
                 "github-copilot/claude-opus-4.1"
                 "github-copilot/claude-opus-4.5"
@@ -51,7 +50,7 @@
     (testing "config updated"
       (is (match?
            {:chat {:models models
-                   :selectModel "anthropic/claude-sonnet-4-5-20250929"
+                   :selectModel "anthropic/claude-sonnet-4.5"
                    :behaviors ["agent" "plan"]
                    :selectBehavior "plan"
                    :welcomeMessage (m/pred #(string/includes? % "Welcome to ECA!"))}}
@@ -75,11 +74,10 @@
 
 (deftest initialize-with-custom-providers
   (eca/start-process!)
-  (let [models ["anthropic/claude-haiku-4-5-20251001"
-                "anthropic/claude-opus-4-1-20250805"
-                "anthropic/claude-opus-4-20250514"
-                "anthropic/claude-sonnet-4-20250514"
-                "anthropic/claude-sonnet-4-5-20250929"
+  (let [models ["anthropic/claude-haiku-4.5"
+                "anthropic/claude-opus-4.1"
+                "anthropic/claude-opus-4.5"
+                "anthropic/claude-sonnet-4.5"
                 "github-copilot/claude-haiku-4.5"
                 "github-copilot/claude-opus-4.1"
                 "github-copilot/claude-opus-4.5"