+++ src/eca/models.clj
@@ -30,10 +30,15 @@
     "openai/gpt-5"
     "openai/gpt-5-mini"
     "openai/gpt-5-nano"
+    "anthropic/claude-sonnet-4.5"
+    "anthropic/claude-opus-4-1"
+    "anthropic/claude-opus-4-5"
+    "anthropic/claude-haiku-4.5"
     "anthropic/claude-sonnet-4-5-20250929"
     "anthropic/claude-sonnet-4-20250514"
     "anthropic/claude-opus-4-20250514"
     "anthropic/claude-opus-4-1-20250805"
+    "anthropic/claude-opus-4-5-20251101"
     "anthropic/claude-haiku-4-5-20251001"})
 
 (defn ^:private all