+++ test/eca/llm_api_test.clj
@@ -30,21 +30,21 @@
                   secrets/credential-file-paths (constantly [])]
       (let [db {:models {}}
             config {:providers {"anthropic" {:key "something"}}}]
-        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config))))))
+        (is (= "anthropic/claude-sonnet-4.5" (llm-api/default-model db config))))))
 
   (testing "Anthropic API key present in ENV"
     (with-redefs [config/get-env (fn [k] (when (= k "ANTHROPIC_API_KEY") "env-anthropic"))
                   secrets/credential-file-paths (constantly [])]
       (let [db {:models {}}
             config {:providers {"anthropic" {:keyEnv "ANTHROPIC_API_KEY"}}}]
-        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config))))))
+        (is (= "anthropic/claude-sonnet-4.5" (llm-api/default-model db config))))))
 
   (testing "OpenAI API key present in config"
     (with-redefs [config/get-env (constantly nil)
                   secrets/credential-file-paths (constantly [])]
       (let [db {:models {}}
             config {:providers {"openai" {:key "yes!"}}}]
-        (is (= "openai/gpt-5" (llm-api/default-model db config))))))
+        (is (= "openai/gpt-5.1" (llm-api/default-model db config))))))
 
   (testing "OpenAI API key present in ENV"
     (with-redefs [config/get-env (fn [k] (when (= k "OPENAI_API_KEY") "env-openai"))
@@ -52,11 +52,11 @@
       (let [db {:models {}}
             config {:providers {"anthropic" {:key nil :keyEnv nil :keyRc nil}
                                 "openai" {:keyEnv "OPENAI_API_KEY"}}}]
-        (is (= "openai/gpt-5" (llm-api/default-model db config))))))
+        (is (= "openai/gpt-5.1" (llm-api/default-model db config))))))
 
   (testing "Fallback default (no keys anywhere)"
     (with-redefs [config/get-env (constantly nil)
                   secrets/credential-file-paths (constantly [])]
       (let [db {:models {}}
             config {}]
-        (is (= "anthropic/claude-sonnet-4-5-20250929" (llm-api/default-model db config)))))))
+        (is (= "anthropic/claude-sonnet-4.5" (llm-api/default-model db config)))))))